backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.31.0
asyncpg==0.31.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
slowapi==0.1.9
openai==1.12.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
